Move backend implementations to `src/backends/`
