+++ src/backends/getentropy.rs
@@ -7,14 +7,17 @@
 //!   - vita newlib since Dec 2021
 //!
 //! For these targets, we use getentropy(2) because getrandom(2) doesn't exist.
-use crate::{util_libc::last_os_error, Error};
+use crate::Error;
 use core::{ffi::c_void, mem::MaybeUninit};
+
+#[path = "../util_libc.rs"]
+mod util_libc;
 
 pub fn fill_inner(dest: &mut [MaybeUninit<u8>]) -> Result<(), Error> {
     for chunk in dest.chunks_mut(256) {
         let ret = unsafe { libc::getentropy(chunk.as_mut_ptr().cast::<c_void>(), chunk.len()) };
         if ret != 0 {
-            return Err(last_os_error());
+            return Err(util_libc::last_os_error());
         }
     }
     Ok(())
