+++ src/backends/vxworks.rs
@@ -1,10 +1,15 @@
 //! Implementation for VxWorks
-use crate::{util_libc::last_os_error, Error};
+use crate::Error;
 use core::{
     cmp::Ordering::{Equal, Greater, Less},
     mem::MaybeUninit,
     sync::atomic::{AtomicBool, Ordering::Relaxed},
 };
+
+#[path = "../util_libc.rs"]
+mod util_libc;
+
+use util_libc::last_os_error;
 
 pub fn fill_inner(dest: &mut [MaybeUninit<u8>]) -> Result<(), Error> {
     static RNG_INIT: AtomicBool = AtomicBool::new(false);
